% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain_packages.R
\name{get_all_loaded_packages}
\alias{get_all_loaded_packages}
\title{Get all loaded packages}
\usage{
get_all_loaded_packages()
}
\value{
Vector with all loaded packages.
}
\description{
Get all loaded packages
}
