% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain_packages.R
\name{get_used_packages}
\alias{get_used_packages}
\title{Get used packages}
\usage{
get_used_packages(used_functions)
}
\arguments{
\item{used_functions}{Vector with used functions.}
}
\value{
Vector with used packages.
}
\description{
Get used packages
}
