% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install missing packages from package list}
\usage{
install_packages(packages)
}
\arguments{
\item{packages}{A vector with package names.}
}
\description{
Install missing packages from package list
}
