% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_environment.R
\name{load_environment}
\alias{load_environment}
\title{Initiates the environment}
\usage{
load_environment(packages, quick = TRUE)
}
\arguments{
\item{packages}{A vector with package names.}

\item{quick}{Logical TRUE to skip prompts.}
}
\description{
Initiates the environment
}
