% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load packages from package lists}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{A vector with package names.}
}
\value{
Vector with manually loaded packages.
}
\description{
Load packages from package lists
}
