% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain_packages.R
\name{maintain_packages}
\alias{maintain_packages}
\title{Maintains the packages}
\usage{
maintain_packages(packageload_man)
}
\arguments{
\item{packageload_man}{Vector with manually loaded packages.}
}
\description{
Maintains the packages
}
