% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{upgrade_packages}
\alias{upgrade_packages}
\title{Upgrades missing packages from package list}
\usage{
upgrade_packages(packages)
}
\arguments{
\item{packages}{A vector with package names.}
}
\description{
Upgrades missing packages from package list
}
