% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{when_updated}
\alias{when_updated}
\title{Return the updated date of locally installed packages}
\usage{
when_updated(pkg, lib = .libPaths()[1])
}
\arguments{
\item{pkg}{The package name}

\item{lib}{}
}
\value{
Returns a POSIXct\link{1:1} format YYYY-MM-DD HH:MM:SS CEST
}
\description{
The function relies on that packages will always have DESCRIPTION file.
According to the author "that's a safe bet" (Konrad, 2021). See
https://stackoverflow.com/questions/68214167/ for details.
}
