% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install missing packages from package list}
\usage{
install_packages(packages, fast = fast)
}
\arguments{
\item{packages}{A vector with package names.}

\item{fast}{Logical TRUE to skip prompts.}

\item{reinstall}{Logical for recursive installation (updating packages)}
}
\description{
Install missing packages from package list
}
